Improve documentation comments on EventBaseManager
Summary:
# What is the goal?

Address the problem in the post at https://fb.workplace.com/groups/code.indexing/permalink/1929080303917865/ with the comment(s) in Codex for EventBaseManager.

# What was not the way?

Jie Jua noticed that the current comment on "EventBaseManager() {}" in Codex was

```
encouraged. You should instead use the global singleton if possible.
```

# What is the new way?

Use documentation comments (doxygen) style `/** double starts */` instead of `//` ordinary comments. This should show the full comment which is:

```
XXX Constructing a EventBaseManager directly is DEPRECATED and not
encouraged. You should instead use the global singleton if possible.
```

Which really needs that first line.

Differential Revision: D24333522

fbshipit-source-id: 8e84e0a49426844d5a6f6438b0d76fb9886742fe