SingleWriterFixedHashMap: Clean up tombstones when copying from a full map
Summary:
Clean up tombstones when copying from a full map.
Equal capacity maps are copied with tombstones for speed but that optimization should not be used if it results in a full map.

Add guarantee that a copy from a map with tombstones will have at least one empty element.

Reviewed By: A5he

Differential Revision: D25230678

fbshipit-source-id: 900b734cc526be8da8be4e70137d19645f994f19