Python error handling
Summary:
This diff introduces a generic abstraction to convert a python error (set by a C-API call) into an informative C++ exception.

The main motivation for this is to improve the error handling behavior when dealing with failed cython api module imports.

This diff adds one callsite to the new abstraction in fiber.cpp  to demonstrate its usage.

Reviewed By: yfeldblum

Differential Revision: D25469594

fbshipit-source-id: e9f83c06460f35fa76030ceae9726297276db850