Expose folly::observer::waitForAllUpdates as public API
Summary:
In most cases, `folly::observer::waitForAllUpdates` should not be needed. However, the asynchronous nature of observers break some basic assumptions that most users expect. Such as:
```
thing.set(value);
EXPECT_EQ(thing.get(), value);
```
If `thing` is using `folly::observer` underneath the hood, the asynchrony is passed on to the user which is very easy to overlook, leading to bugs. The API of `thing` might want to enforce that when `thing.set` returns, all observed values will have been updated.
Blocking is a perfectly okay solution in cases we don't expect an observer to be updated frequently.

Differential Revision: D25542376

fbshipit-source-id: c83909807fbf36e0311a45b667da94c43751a1da