allow specifying multiple CPUs in io_uring SQ_POLL mode.
Summary:
Existing API allows user to specify a single preferred CPU for SQ_POLL workers,
but it also allows specifying multiple workers.

Add API to allow specifying a set of CPUs, not just one.

The specified set of CPUs are used in round-robin fashion, so it is not strictly
necessary that the number of CPUs and the number of workers be the same.

#forceTDHashing

Reviewed By: yfeldblum

Differential Revision: D26380317

fbshipit-source-id: ecfd8722670ffc3e3dc246fae1bcb7f4d8e6f3e4